# openapi.yml
openapi: 3.0.3
info:
  title: AHB-Tabellen API Documentation
  description: >
    This is the description for the API of the ahb-tabellen web application.

    Some useful links:
    - [The machine-readable-anwendungshandbuecher repository](https://github.com/Hochfrequenz/machine-readable_anwendungshandbuecher)
    - [kohlr_AHB_i - the CLI tool to scrape the AHB tables from docx files](https://github.com/Hochfrequenz/kohlrahbi)
  termsOfService: http://swagger.io/terms/
  contact:
    email: kevin.krechan@hochfrequenz.de
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 0.0.1
...
servers:
  - url: https://ahb-tabellen.hochfrequenz.de/api/v1
...
paths:
...
  /api/pruefi-diff/{pruefiOld}/{pruefiNew}:
    get:
      tags:
        - ahb
      summary: Get a diff comparison between two different Pruefidentifikatoren within the same format version
      operationId: getPruefiDiff
      parameters:
        - name: pruefiOld
          in: path
          description: First Pruefidentifikator to compare
          required: true
          schema:
            type: string
            pattern: '^\d{5}$'
            example: '13002'
        - name: pruefiNew
          in: path
          description: Second Pruefidentifikator to compare
          required: true
          schema:
            type: string
            pattern: '^\d{5}$'
            example: '13003'
        - name: format-version
          in: query
          description: The format version both Pruefidentifikatoren are compared in (FVxxxx code, 'current', or an ISO date YYYY-MM-DD)
          required: true
          schema:
            $ref: '#/components/schemas/FormatVersion'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AhbPruefiDiff'
        '400':
          description: pruefiOld and pruefiNew are identical
        '404':
          description: One or both Pruefidentifikatoren not found in the given format version
...
components:
  schemas:
...
    FormatVersion:
      type: string
      description: >
        An EDIFACT format version code (`FVYYYY`, e.g. `FV2510`), the keyword `current`, or an
        ISO date (`YYYY-MM-DD`). `current` is resolved server-side to the version valid right now
        and a date to the version valid on that day.
      example: 'FV2510'
...
    AhbPruefiDiff:
      type: object
      description: >
        Diff comparison between two different Pruefidentifikatoren within the same format version
        (as opposed to AhbDiff, which compares the same Pruefidentifikator across two format versions)
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/AhbDiffLine'
        meta:
          type: object
          properties:
            format_version:
              type: string
              example: 'FV2410'
            pruefidentifikator_old:
              type: string
              example: '13002'
            pruefidentifikator_new:
              type: string
              example: '13003'
            description_new:
              type: string
              nullable: true
              description: Description of the new (second) Pruefidentifikator
            description_old:
              type: string
              nullable: true
              description: Description of the old (first) Pruefidentifikator
          required:
            - format_version
            - pruefidentifikator_old
            - pruefidentifikator_new
      required:
        - lines
        - meta
    AhbDiffLine:
      type: object
      description: A line in the AHB diff comparison, containing data from both versions joined from v_ahbtabellen
      properties:
        diff_status:
          type: string
          enum: [unchanged, modified, added, deleted]
          description: The diff status from v_ahb_diff
        id_path:
          type: string
          description: The id_path from v_ahb_diff used to join with v_ahbtabellen
        sort_path:
          type: string
          description: The sort_path for ordering
        changed_columns:
          type: array
          items:
            type: string
          description: List of column names that changed between versions
        old:
          $ref: '#/components/schemas/AhbDiffSide'
          description: Data from v_ahbtabellen for the old (B) version
          nullable: true
        new:
          $ref: '#/components/schemas/AhbDiffSide'
          description: Data from v_ahbtabellen for the new (A) version
          nullable: true
      required:
        - diff_status
        - id_path
        - sort_path
    AhbDiffSide:
      type: object
      description: AHB line data from v_ahbtabellen for one side of the comparison
      properties:
        segmentgroup_key:
          type: string
          nullable: true
        segment_code:
          type: string
          nullable: true
        data_element:
          type: string
          nullable: true
        qualifier:
          type: string
          nullable: true
        line_ahb_status:
          type: string
          nullable: true
        line_name:
          type: string
          nullable: true
        line_type:
          type: string
          nullable: true
        bedingung:
          type: string
          nullable: true
        ebd_key:
          type: string
          nullable: true
          description: EBD decision-tree key (e.g. E_0401) detected in qualifier, if any
...